Extract common actor and time properties
