+++ core/src/main/java/io/spine/core/CommandContextMixin.java
@@ -21,41 +21,17 @@
 package io.spine.core;
 
 import com.google.errorprone.annotations.Immutable;
-import com.google.protobuf.Message;
-import io.spine.base.MessageContext;
-import io.spine.core.Enrichment.Container;
-
-import java.util.Optional;
-
-import static io.spine.core.Enrichments.containerIn;
+import io.spine.annotation.GeneratedMixin;
 
 /**
- * A common interface for message contexts that hold enrichments.
+ * Mix-in interface extending {@link CommandContext}.
  */
+@GeneratedMixin
 @Immutable
-public interface EnrichableMessageContext extends MessageContext {
+public interface CommandContextMixin extends CommandContextOrBuilder, ContextWithActor {
 
-    /**
-     * Obtains an instance of {@link Enrichment} from the context of the message.
-     */
-    @SuppressWarnings("override") // in generated code.
-    Enrichment getEnrichment();
-
-    default <E extends Message> Optional<E> find(Class<E> cls) {
-        Optional<Container> container = containerIn(this);
-        Optional<E> result = container.flatMap(c -> Enrichments.find(cls, c));
-        return result;
-    }
-
-    /**
-     * Obtains enrichment of the passed class.
-     *
-     * @throws IllegalStateException if the enrichment is not found
-     */
-    default <E extends Message> E get(Class<E> cls) {
-        Container container = containerIn(this).orElse(Container.getDefaultInstance());
-        E result = Enrichments.get(cls, container);
-        return result;
+    @Override
+    default ActorContext actorContext() {
+        return getActorContext();
     }
 }
-
