+++ core/src/main/java/io/spine/core/ContextWithActor.java
@@ -21,41 +21,11 @@
 package io.spine.core;
 
 import com.google.errorprone.annotations.Immutable;
-import com.google.protobuf.Message;
 import io.spine.base.MessageContext;
-import io.spine.core.Enrichment.Container;
-
-import java.util.Optional;
-
-import static io.spine.core.Enrichments.containerIn;
 
 /**
- * A common interface for message contexts that hold enrichments.
+ * A context of a message initiated by a user.
  */
 @Immutable
-public interface EnrichableMessageContext extends MessageContext {
-
-    /**
-     * Obtains an instance of {@link Enrichment} from the context of the message.
-     */
-    @SuppressWarnings("override") // in generated code.
-    Enrichment getEnrichment();
-
-    default <E extends Message> Optional<E> find(Class<E> cls) {
-        Optional<Container> container = containerIn(this);
-        Optional<E> result = container.flatMap(c -> Enrichments.find(cls, c));
-        return result;
-    }
-
-    /**
-     * Obtains enrichment of the passed class.
-     *
-     * @throws IllegalStateException if the enrichment is not found
-     */
-    default <E extends Message> E get(Class<E> cls) {
-        Container container = containerIn(this).orElse(Container.getDefaultInstance());
-        E result = Enrichments.get(cls, container);
-        return result;
-    }
+public interface ContextWithActor extends MessageContext, WithActor {
 }
-
